Pull down upstream changes to qonsole-rails
In pursuit of reducing the number of spurious Sentry error messages we
get, pull down the upstream changes to qonsole-rails that fix a path
issue with rendering error pages.

Also allow the relative_url_path to be overridden from the environment,
in service of better testing the app in production mode.