WIP: working on converting to use webpacker
This is a WIP commit for partial progess in changing to use webpacker
for the lr-landing site. The core problem is that many of the JS and CSS
assets we rely are currently packaged as Rails assets gems. This is
mostly deprecated now, and specifically some packages (e.g.
codemirror-rails) will not be updated for Rails 6. Therefore, in order
to move to Rails 6 we'll need to move *all* of the JS and CSS
dependencies to webpacker - this includes the common styles, and
Qonsole. It's potentially quite a big job.