Add Welsh-localized accessibility statement
The core objective in this commit was to add the translated
accessibility statement. However, it was clear once I had done so that
it made no sense to localize the content of the statement without also
updating the header menu and footer, and adding a switcher so that users
can swap between English and Welsh. This is now done, mostly via updates
to the shared-styles gem.

I've also switched to a common pattern for pages that have two versions:
if we want to localize `page.html.haml` we create partial versions
`page_en.html.haml`, `page_cy.html.haml`, etc, and then the main
`page.html.haml` template simply has to render the right partial based
on the current value of `I18n.locale`.